Change v256_ssd_s16_sum() to return uint64_t.
Fix the Visual Studio 2015 and 2017 compiler warning:
  aom_dsp\simd/v256_intrinsics.h(93,31): warning C4244: 'return':
  conversion from 'uint64_t' to 'uint32_t', possible loss of data

Evidences that v256_ssd_s16_sum() should return uint64_t:

1. The other two implementations of this function both return uint64_t.

  aom_dsp/simd/v256_intrinsics_v128.h:
  SIMD_INLINE uint64_t v256_ssd_s16_sum(ssd256_internal_s16 s) {
    return v128_ssd_s16_sum(s.val[1]) + v128_ssd_s16_sum(s.val[0]);
  }

  aom_dsp/simd/v256_intrinsics_x86.h:
  SIMD_INLINE uint64_t v256_ssd_s16_sum(ssd256_internal_s16 s) {
    v128 t = v128_add_64(v256_high_v128(s), v256_low_v128(s));
    return v64_u64(v128_low_v64(t)) + v64_u64(v128_high_v64(t));
  }

2. The unit test apparently expects this function to return uint64_t.

  test/simd_cmp_impl.h:
  uint64_t v256_ssd_s16(v256 a, v256 b) {
    return v256_ssd_s16_sum(::v256_ssd_s16(v256_ssd_s16_init(), a, b));
  }

+++ aom_dsp/simd/v256_intrinsics.h
@@ -89,7 +89,7 @@ SIMD_INLINE c_ssd256_internal_s16 v256_ssd_s16(c_ssd256_internal_s16 s, v256 a,
                                                v256 b) {
   return c_v256_ssd_s16(s, a, b);
 }
-SIMD_INLINE uint32_t v256_ssd_s16_sum(c_ssd256_internal_s16 s) {
+SIMD_INLINE uint64_t v256_ssd_s16_sum(c_ssd256_internal_s16 s) {
   return c_v256_ssd_s16_sum(s);
 }
